requirements.txt
argcomplete
Biopython
pandas
numpy
selenium
cutadapt
bowtie
openpyxl
MACS3
